rename all scene_part files
