+++ crates/re_space_view_bar_chart/src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! A Space View that shows a single bar chart.
 
-mod scene_part;
 mod space_view_class;
+mod view_part_system;
 
 pub use space_view_class::BarChartSpaceView;
