+++ crates/re_space_view_bar_chart/src/view_part_system.rs
