+++ crates/re_space_view_tensor/src/lib.rs
@@ -2,9 +2,9 @@
 //!
 //! A Space View dedicated to visualizing tensors with arbitrary dimensionality.
 
-mod scene_part;
 mod space_view_class;
 mod tensor_dimension_mapper;
 mod tensor_slice_to_gpu;
+mod view_part_system;
 
 pub use space_view_class::TensorSpaceView;
