+++ crates/re_space_view_tensor/src/view_part_system.rs
