+++ crates/re_space_view_text/src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! A Space View that shows text entries in a table and scrolls with the active time.
 
-mod scene_part;
 mod space_view_class;
+mod view_part_system;
 
 pub use space_view_class::TextSpaceView;
