+++ crates/re_space_view_text_box/src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! A simple Space View that shows a single text box.
 
-mod scene_part;
 mod space_view_class;
+mod view_part_system;
 
 pub use space_view_class::TextBoxSpaceView;
