+++ crates/re_space_view_time_series/src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! A Space View that shows plots over Rerun timelines.
 
-mod scene_part;
 mod space_view_class;
+mod view_part_system;
 
 pub use space_view_class::TimeSeriesSpaceView;
