+++ crates/re_space_view_time_series/src/space_view_class.rs
@@ -11,7 +11,7 @@ use re_viewer_context::{
     SpaceViewClass, SpaceViewClassName, SpaceViewId, TypedScene, ViewerContext,
 };
 
-use crate::scene_part::{PlotSeriesKind, SceneTimeSeries};
+use crate::view_part_system::{PlotSeriesKind, SceneTimeSeries};
 
 #[derive(Default)]
 pub struct TimeSeriesSpaceView;
