+++ crates/re_viewer_context/src/space_view/mod.rs
@@ -8,11 +8,11 @@ mod dyn_space_view_class;
 mod highlights;
 mod scene;
 mod scene_context;
-mod scene_part;
 mod scene_query;
 mod space_view_class;
 mod space_view_class_placeholder;
 mod space_view_class_registry;
+mod view_part_system;
 
 pub use dyn_space_view_class::{
     ArchetypeDefinition, DynSpaceViewClass, SpaceViewClassLayoutPriority, SpaceViewClassName,
@@ -21,7 +21,7 @@ pub use dyn_space_view_class::{
 pub use highlights::{SpaceViewEntityHighlight, SpaceViewHighlights, SpaceViewOutlineMasks};
 pub use scene::{Scene, TypedScene};
 pub use scene_context::{SceneContext, SceneContextPart};
-pub use scene_part::{ScenePartCollection, ViewPartSystem};
 pub use scene_query::SceneQuery;
 pub use space_view_class::SpaceViewClass;
 pub use space_view_class_registry::{SpaceViewClassRegistry, SpaceViewClassRegistryError};
+pub use view_part_system::{ScenePartCollection, ViewPartSystem};
