+++ crates/re_viewer_context/src/space_view/view_part_system.rs
