+++ examples/rust/custom_space_view/src/color_coordinates_view_part_system.rs
